remove specs
I prefer to actually start this exercise without tests. Choosing which
tests to write, and which tests to write *first* are both important
decisions that are worthy of discussion.

Since I like to do this problem as a pair, it makes sense to leave out
code that sort of makes these decisions for you.